⚡ Bolt: [Optimize NumPy Array Memory Footprint for Embeddings]

+++ rag_engine.py
@@ -95,16 +95,16 @@ class RAGEngine:
                         candidates.append((pair["raw"], pair["translated"], emb))
 
         if not candidates:
-            self._cached_tm_matrix = np.array([])
-            self._cached_tm_norms = np.array([])
+            self._cached_tm_matrix = np.array([], dtype=np.float32)
+            self._cached_tm_norms = np.array([], dtype=np.float32)
             self._cached_candidates = []
             return
 
         self._cached_candidates = candidates
 
         # Create matrix and precalculate norms
-        self._cached_tm_matrix = np.array([emb for _, _, emb in candidates])
-        self._cached_tm_norms = np.linalg.norm(self._cached_tm_matrix, axis=1)
+        self._cached_tm_matrix = np.array([emb for _, _, emb in candidates], dtype=np.float32)
+        self._cached_tm_norms = np.linalg.norm(self._cached_tm_matrix, axis=1).astype(np.float32)
         # Avoid division by zero
         self._cached_tm_norms[self._cached_tm_norms == 0] = 1e-9
 
@@ -156,7 +156,7 @@ class RAGEngine:
         if len(self._cached_candidates) == 0:
             return []
 
-        q_arr = np.array(q_emb)
+        q_arr = np.array(q_emb, dtype=np.float32)
         q_norm = np.linalg.norm(q_arr) or 1e-9
 
         sims = np.dot(self._cached_tm_matrix, q_arr) / (self._cached_tm_norms * q_norm)
@@ -313,7 +313,7 @@ class RAGEngine:
             emb = pair.get("embedding")
             if emb and isinstance(emb, list) and len(emb) > 0:
                 vectors.append(emb)
-        result = np.mean(vectors, axis=0) if vectors else None
+        result = np.mean(vectors, axis=0, dtype=np.float32) if vectors else None
         self._chapter_tm_embeddings_cache[filename] = result
         return result
 
@@ -347,7 +347,7 @@ class RAGEngine:
             paras = [p.strip() for p in curr_text.split("\n\n") if p.strip()][:3]
             if paras:
                 curr_embs = [self._generate_embedding_sync(p) for p in paras]
-                curr_emb = np.mean(curr_embs, axis=0)
+                curr_emb = np.mean(curr_embs, axis=0, dtype=np.float32)
                 best_chap = self._find_best_semantic_match(curr_emb, candidates)
                 if best_chap is not None:
                     return chapter_dict[best_chap]
